add Route to App.js
